requirements.txt
alembic==1.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
et-xmlfile==1.1.0
filelock==3.9.0
Flask==2.2.2
Flask-Dance==6.2.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
huggingface-hub==0.11.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.1
numpy==1.24.1
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
pinecone-client==2.0.13
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.45
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.25.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
URLObject==2.4.3
Werkzeug==2.2.2
